style-105: Code Clean Up
