+++ data/src/main/java/com/plub/data/api/PlubJwtTokenApi.kt
@@ -1,7 +1,7 @@
 package com.plub.data.api
 
 import com.plub.data.model.PlubJwtTokenResponse
-import com.plub.data.model.ReIssueModel
+import com.plub.data.model.JWTTokenReIssueRequest
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
@@ -9,5 +9,5 @@ import retrofit2.http.POST
 interface PlubJwtTokenApi {
     @POST("/api/auth/reissue/")
     suspend fun reIssueToken(
-        @Body refreshToken : ReIssueModel) : Response<PlubJwtTokenResponse>
+        @Body refreshToken : JWTTokenReIssueRequest) : Response<PlubJwtTokenResponse>
 }
