+++ data/src/main/java/com/plub/data/model/JWTTokenReIssueRequest.kt
@@ -1,5 +1,5 @@
 package com.plub.data.model
 
-data class ReIssueModel(
+data class JWTTokenReIssueRequest(
     val refreshToken : String
 )
