--- data/src/main/java/com/plub/data/model/PlubJwtTokenData.kt
@@ -1,11 +1,0 @@
-package com.plub.data.model
-
-import com.google.gson.annotations.SerializedName
-
-data class PlubJwtTokenData(
-    @SerializedName("accessToken")
-    val accessToken : String,
-
-    @SerializedName("refreshToken")
-    val refreshToken : String
-)
